One refactor more in the binding, metadata and related stuff
