+++ xmipp3/protocols/protocol_solid_angles.py
@@ -37,7 +37,7 @@ from pwem.protocols import ProtAnalysis3D
 import pwem.emlib.metadata as md
 
 from pwem import emlib
-from xmipp3.base import findRow
+from xmipp3.base import findRow, readInfoField, writeInfoField
 from xmipp3.convert import (rowToAlignment, setXmippAttributes, xmippToLocation,
                             createItemMatrix, writeSetOfParticles)
 from xmipp3.constants import SYM_URL
@@ -165,17 +165,6 @@ class XmippProtSolidAngles(ProtAnalysis3D):
         self._insertFunctionStep('createOutputStep')
     
     #--------------------------- STEPS functions -------------------------------
-
-    def readInfoField(self,fnDir,block,label):
-        mdInfo = emlib.MetaData("%s@%s"%(block,join(fnDir,"iterInfo.xmd")))
-        return mdInfo.getValue(label,mdInfo.firstObject())
-
-    def writeInfoField(self,fnDir,block,label, value):
-        mdInfo = emlib.MetaData()
-        objId=mdInfo.addObject()
-        mdInfo.setValue(label,value,objId)
-        mdInfo.write("%s@%s"%(block,join(fnDir,"iterInfo.xmd")),emlib.MD_APPEND)
-
     def convertInputStep(self, particlesId, volId):
         """ Write the input images as a Xmipp metadata file. 
         particlesId: is only need to detect changes in
@@ -210,8 +199,8 @@ class XmippProtSolidAngles(ProtAnalysis3D):
             
             Xdim=newXdim
             Ts=newTs
-        self.writeInfoField(self._getExtraPath(),"sampling",emlib.MDL_SAMPLINGRATE,Ts)
-        self.writeInfoField(self._getExtraPath(),"size",emlib.MDL_XSIZE,int(Xdim))
+        writeInfoField(self._getExtraPath(),"sampling",emlib.MDL_SAMPLINGRATE,Ts)
+        writeInfoField(self._getExtraPath(),"size",emlib.MDL_XSIZE,int(Xdim))
 
     def constructGroupsStep(self, particlesId, angularSampling,
                             angularDistance, symmetryGroup):
@@ -389,8 +378,8 @@ class XmippProtSolidAngles(ProtAnalysis3D):
         fnTmpDir = self._getTmpPath()
         fnDirectional = self._getDirectionalClassesFn()
         inputParticles = self.inputParticles.get()
-        newTs = self.readInfoField(self._getExtraPath(),"sampling",emlib.MDL_SAMPLINGRATE)
-        newXdim = self.readInfoField(self._getExtraPath(),"size",emlib.MDL_XSIZE)
+        newTs = readInfoField(self._getExtraPath(),"sampling",emlib.MDL_SAMPLINGRATE)
+        newXdim = readInfoField(self._getExtraPath(),"size",emlib.MDL_XSIZE)
 
         # Generate projections
         fnGallery=join(fnTmpDir,"gallery.stk")
@@ -422,8 +411,8 @@ class XmippProtSolidAngles(ProtAnalysis3D):
         cleanPattern(self._getExtraPath("direction_*"))
 
     def splitVolumeStep(self):
-        newTs = self.readInfoField(self._getExtraPath(),"sampling",emlib.MDL_SAMPLINGRATE)
-        newXdim = self.readInfoField(self._getExtraPath(),"size",emlib.MDL_XSIZE)
+        newTs = readInfoField(self._getExtraPath(),"sampling",emlib.MDL_SAMPLINGRATE)
+        newXdim = readInfoField(self._getExtraPath(),"size",emlib.MDL_XSIZE)
         fnMask = ""
         if self.mask.hasValue():
             fnMask = self._getExtraPath("mask.vol")
@@ -444,7 +433,7 @@ class XmippProtSolidAngles(ProtAnalysis3D):
         if not self._useSeveralClasses():
             newTs = inputParticles.getSamplingRate()
         else:
-            newTs = self.readInfoField(self._getExtraPath(),"sampling",emlib.MDL_SAMPLINGRATE)
+            newTs = readInfoField(self._getExtraPath(),"sampling",emlib.MDL_SAMPLINGRATE)
 
         self.mdClasses = emlib.MetaData(self._getDirectionalClassesFn())
         self.mdImages = emlib.MetaData(self._getDirectionalImagesFn())
